requirements.txt
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
celery==4.0.2
ciso8601==1.0.3
click==6.7
Django==1.10.5
django-cal==0.2.3
kombu==4.0.2
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==1.5
python-dotenv==0.6.3
pytz==2016.10
redis==2.10.5
requests==2.13.0
six==1.10.0
vine==1.1.3
vobject==0.9.4.1
